refactor(General): :recycle: Rename and move config files
