+++ cypress/e2e/skupper-console/shared/navigation.cy.ts
@@ -1,4 +1,4 @@
-import { getTestsIds } from '@config/testIds.config';
+import { getTestsIds } from '@config/testIds';
 
 context('Navigation', () => {
   beforeEach(() => {
