+++ src/core/components/SkBreadcrumb/index.tsx
@@ -1,7 +1,7 @@
 import { Breadcrumb, BreadcrumbHeading, BreadcrumbItem } from '@patternfly/react-core';
 import { Link, useLocation } from 'react-router-dom';
 
-import { getTestsIds } from '@config/testIds.config';
+import { getTestsIds } from '@config/testIds';
 import { getIdAndNameFromUrlParams } from '@core/utils/getIdAndNameFromUrlParams';
 
 const SkBreadcrumb = function () {
